# api/pki-ca/openapi_spec_enrollment.yaml
...
openapi: 3.0.3
info:
  title: CA Enrollment API
  version: "1.0.0"
  description: >
    Certificate Authority Enrollment API for certificate request submission,
    issuance, retrieval, and status checking. Supports multiple certificate
    profiles including TLS server, TLS client, code signing, and S/MIME.
  contact:
    name: API Support
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: /api/v1/ca
    description: CA API base path
...
paths:
...
  /ocsp:
    post:
      tags:
        - OCSP
      summary: OCSP responder endpoint
      description: >
        Online Certificate Status Protocol (OCSP) responder endpoint per RFC 6960.
        Accepts OCSP requests and returns signed OCSP responses indicating the
        revocation status of certificates.
      operationId: handleOCSP
      requestBody:
        required: true
        content:
          application/ocsp-request:
            schema:
              type: string
              format: binary
              description: DER-encoded OCSP request
      responses:
        '200':
          description: OCSP response
          content:
            application/ocsp-response:
              schema:
                type: string
                format: binary
                description: DER-encoded OCSP response
        '400':
          description: Invalid OCSP request
          content:
            application/ocsp-response:
              schema:
                type: string
                format: binary
        '500':
          description: Internal server error
          content:
            application/ocsp-response:
              schema:
                type: string
                format: binary

  # EST Protocol Endpoints (RFC 7030)
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT authentication token
security:
  - bearerAuth: []